platform dependent prop parse added
